Create a Hangfire folder under the web project with a Jobs and MediatR folder
- the jobs folder holds jobs that are scheduled by Hangfire
- the MediatR folder holds messages/handlers that are sent by Hangfire jobs
