+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Features/Notifications/RequestConfirmationsSentHandler.cs
@@ -1,7 +1,7 @@
 ﻿using System.Linq;
 using System.Threading.Tasks;
-using AllReady.Areas.Admin.RequestConfirmationMessageSenders;
 using AllReady.Extensions;
+using AllReady.Hangfire.Jobs;
 using AllReady.Models;
 using Hangfire;
 using MediatR;
