+++ AllReadyApp/Web-App/AllReady/Hangfire/Jobs/DayOfRequestConfirmationMessageSender.cs
@@ -1,38 +1,36 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using AllReady.Extensions;
+using AllReady.Hangfire.MediatR;
 using AllReady.Models;
 using AllReady.Models.Notifications;
 using AllReady.Services;
-using Hangfire;
+using MediatR;
 using Newtonsoft.Json;
 
-namespace AllReady.Areas.Admin.RequestConfirmationMessageSenders
+namespace AllReady.Hangfire.Jobs
 {
-    public class DayBeforeRequestConfirmationMessageSender : IDayBeforeRequestConfirmationMessageSender
+    public class DayOfRequestConfirmationMessageSender : IDayOfRequestConfirmationMessageSender
     {
         private readonly AllReadyContext context;
         private readonly IQueueStorageService storageService;
-        private readonly IBackgroundJobClient backgroundJob;
+        private readonly IMediator mediator;
+
         public Func<DateTime> DateTimeUtcNow = () => DateTime.UtcNow;
 
-        public DayBeforeRequestConfirmationMessageSender(AllReadyContext context, IQueueStorageService storageService, IBackgroundJobClient backgroundJob)
+        public DayOfRequestConfirmationMessageSender(AllReadyContext context, IQueueStorageService storageService, IMediator mediator)
         {
             this.context = context;
             this.storageService = storageService;
-            this.backgroundJob = backgroundJob;
+            this.mediator = mediator;
         }
 
         public void SendSms(List<Guid> requestIds, int itineraryId)
         {
             var requests = context.Requests.Where(x => requestIds.Contains(x.RequestId) && x.Status == RequestStatus.PendingConfirmation).ToList();
-            var itinerary = context.Itineraries.Single(x => x.Id == itineraryId);
+            var itineraryDate = context.Itineraries.Single(x => x.Id == itineraryId).Date;
 
-            //don't send out messages if DateTime.UtcNow is less than 1 days away from the Itinerary.Date. 
-            //This can happen if a request is added to an itinereary less than 7 days away from the itinerary's date
-            //Hangfire will execute the job immediately when the scheduled time has already passed
-            if (DateTimeUtcNow().Date >= itinerary.Date.AddDays(-1).Date)
+            if (DateTimeUtcNow().Date == itineraryDate.Date)
             {
                 //TODO mgmccarthy: need to convert intinerary.Date to local time of the request's intinerary's campaign's timezone
                 requests.ForEach(request =>
@@ -40,18 +38,18 @@ namespace AllReady.Areas.Admin.RequestConfirmationMessageSenders
                     var queuedSms = new QueuedSmsMessage
                     {
                         Recipient = request.Phone,
-                        Message = $@"Your request has been scheduled by allReady for {itinerary.Date}. Please response with ""Y"" to confirm this request or ""N"" to cancel this request."
+                        Message = "sorry you couldn't make it, we will reschedule."
                     };
                     var sms = JsonConvert.SerializeObject(queuedSms);
                     storageService.SendMessageAsync(QueueStorageService.Queues.SmsQueue, sms);
                 });
             }
             
-            backgroundJob.Schedule<IDayOfRequestConfirmationMessageSender>(x => x.SendSms(requestIds, itinerary.Id), itinerary.Date.AtNoon());
+            mediator.Send(new SetRequstsToUnassignedCommand { RequestIds = requests.Select(x => x.RequestId).ToList() });
         }
     }
 
-    public interface IDayBeforeRequestConfirmationMessageSender
+    public interface IDayOfRequestConfirmationMessageSender
     {
         void SendSms(List<Guid> requestIds, int itineraryId);
     }
