+++ AllReadyApp/Web-App/AllReady/Hangfire/MediatR/SetRequstsToUnassignedCommand.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using MediatR;
 
-namespace AllReady.Areas.Admin.Features.Requests
+namespace AllReady.Hangfire.MediatR
 {
     public class SetRequstsToUnassignedCommand : IRequest
     {
